Remove pipe in apply script (#8)
* Remove pipe to in apply script

This was a failing the shellcheck linting with the message: SC2015: Note that A && B || C is not if-then-else. C may run when A is true. I decided to remove it as each script has its own exit condition.

* Remove branch filter in circle config

The branch filtering was messing with the configuration, so I have removed it.